.Wav File Cutter
This code cuts .wav files to the ASA/ANSI standard of the duration the time it takes for the ship to travel 1 length about the closest point of approach. The code requires that you run Sean's ship code in order to use the .txt and .wav file outputs. The code requires that you have a separate .xlsx file with all of the ship length measurements. It could be modified if you have accurate ship length measurements from AIS data.