Spherical Spreading TL Frequency Dependent
This is a function that calculates frequency dependent spherical spreading transmission loss between a source and a receiver. The function inputs are the start and end frequency, depth of receiver, depth of source, horizontal range from source to receiver, pH, temperature, salinity, and sound speed. The function output is the frequency dependent spherical spreading transmission loss for the frequencies under investigation.